+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 0.1.9 (unreleased)
 ++++++++++++++++++
 
-- Nothing changed yet.
+- Backport the `sqlall` command.
 
 
 0.1.8 (2017-09-20)